finish redis service

+++ srcs/requirements/bonus/StaticWebsite/tools/setup-static-website.sh
@@ -1,3 +1,7 @@
 #!/bin/bash
-mkdir -p /var/www/html/bnaji 
-mv /tmp/portfolio/* /var/www/html/bnaji
+if [ ! -f /tmp/portfolio ];
+then
+  mkdir -p /var/www/html/bnaji 
+  mv /tmp/portfolio/* /var/www/html/bnaji
+  rm -rf /tmp/portfolio
+fi
